Group mass attenuation and mass energy-absorption
